+++ __test__/dom.spec.ts
@@ -1,7 +1,0 @@
-import { getOffset } from '../src/dom'
-
-describe('dom', () => {
-  test('getOffset', () => {
-    
-  })
-})